Coder reorganization
HuffmanCoder was getting huge, so split the encoder and decoder into two files
and interfaces.
