+++ fly/coders/huffman/huffman_encoder.cpp
@@ -1,11 +1,10 @@
-#include "fly/coders/huffman_coder.h"
+#include "fly/coders/huffman/huffman_encoder.h"
 
 #include "fly/coders/bit_stream.h"
 #include "fly/literals.h"
 
 #include <algorithm>
 #include <limits>
-#include <memory>
 #include <stack>
 #include <vector>
 
@@ -20,13 +19,14 @@ namespace {
 
     constexpr const length_type s_maxCodeLength = 11;
 
-    static_assert(std::numeric_limits<code_type>::digits >= s_maxCodeLength,
+    static_assert(
+        std::numeric_limits<code_type>::digits >= s_maxCodeLength,
         "Maximum Huffman code length is too large for code_type");
 
 } // namespace
 
 //==============================================================================
-bool HuffmanCoder::EncodeInternal(
+bool HuffmanEncoder::EncodeInternal(
     std::istream &input,
     BitStreamWriter &output) noexcept
 {
@@ -62,38 +62,7 @@ bool HuffmanCoder::EncodeInternal(
 }
 
 //==============================================================================
-bool HuffmanCoder::DecodeInternal(
-    BitStreamReader &input,
-    std::ostream &output) noexcept
-{
-    std::uint32_t chunkSize;
-    length_type maxCodeLength;
-
-    if (!decodeHeader(input, chunkSize, maxCodeLength))
-    {
-        return false;
-    }
-
-    m_chunkBuffer = std::make_unique<symbol_type[]>(chunkSize);
-    m_prefixTable = std::make_unique<HuffmanCode[]>(1_u64 << maxCodeLength);
-
-    while (!input.FullyConsumed())
-    {
-        if (!decodeCodes(input, maxCodeLength))
-        {
-            return false;
-        }
-        else if (!decodeSymbols(input, maxCodeLength, chunkSize, output))
-        {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-//==============================================================================
-std::uint32_t HuffmanCoder::readStream(std::istream &input) const noexcept
+std::uint32_t HuffmanEncoder::readStream(std::istream &input) const noexcept
 {
     const std::istream::pos_type start = input.tellg();
     input.seekg(0, std::ios::end);
@@ -114,7 +83,7 @@ std::uint32_t HuffmanCoder::readStream(std::istream &input) const noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::createTree(std::uint32_t chunkSize) noexcept
+bool HuffmanEncoder::createTree(std::uint32_t chunkSize) noexcept
 {
     // Lambda to retrieve the next available HuffmanNode
     std::uint16_t index = 0;
@@ -186,7 +155,7 @@ bool HuffmanCoder::createTree(std::uint32_t chunkSize) noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::createCodes() noexcept
+bool HuffmanEncoder::createCodes() noexcept
 {
     std::stack<const HuffmanNode *> pending;
     std::stack<const HuffmanNode *> path;
@@ -245,7 +214,7 @@ bool HuffmanCoder::createCodes() noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::insertCode(HuffmanCode &&code) noexcept
+bool HuffmanEncoder::insertCode(HuffmanCode &&code) noexcept
 {
     if (m_huffmanCodesSize == m_huffmanCodes.size())
     {
@@ -264,7 +233,7 @@ bool HuffmanCoder::insertCode(HuffmanCode &&code) noexcept
 }
 
 //==============================================================================
-void HuffmanCoder::limitCodeLengths() noexcept
+void HuffmanEncoder::limitCodeLengths() noexcept
 {
     auto computeKraft = [](const HuffmanCode &code) -> code_type {
         return 1_u16 << (s_maxCodeLength - code.m_length);
@@ -316,7 +285,7 @@ void HuffmanCoder::limitCodeLengths() noexcept
 }
 
 //==============================================================================
-void HuffmanCoder::convertToCanonicalForm() noexcept
+void HuffmanEncoder::convertToCanonicalForm() noexcept
 {
     // First code is always set to zero. Its length does not change.
     m_huffmanCodes[0].m_code = 0;
@@ -342,7 +311,7 @@ void HuffmanCoder::convertToCanonicalForm() noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::encodeHeader(BitStreamWriter &output) const noexcept
+bool HuffmanEncoder::encodeHeader(BitStreamWriter &output) const noexcept
 {
     // Encode the Huffman coder version.
     if (!output.WriteByte(static_cast<byte_type>(s_huffmanVersion)))
@@ -366,59 +335,7 @@ bool HuffmanCoder::encodeHeader(BitStreamWriter &output) const noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::decodeHeader(
-    BitStreamReader &input,
-    std::uint32_t &chunkSize,
-    length_type &maxCodeLength) const noexcept
-{
-    // Decode the Huffman coder version.
-    byte_type huffmanVersion;
-
-    if (!input.ReadByte(huffmanVersion))
-    {
-        return false;
-    }
-
-    switch (huffmanVersion)
-    {
-        case 1:
-            return decodeHeaderVersion1(input, chunkSize, maxCodeLength);
-
-        default:
-            break;
-    }
-
-    return false;
-}
-
-//==============================================================================
-bool HuffmanCoder::decodeHeaderVersion1(
-    BitStreamReader &input,
-    std::uint32_t &chunkSize,
-    length_type &maxCodeLength) const noexcept
-{
-    // Decode the chunk size.
-    word_type encodedChunkSizeKB;
-    if (!input.ReadWord(encodedChunkSizeKB))
-    {
-        return false;
-    }
-
-    // Decode the maximum Huffman code length.
-    byte_type encodedMaxCodeLength;
-    if (!input.ReadByte(encodedMaxCodeLength))
-    {
-        return false;
-    }
-
-    chunkSize = static_cast<std::uint32_t>(encodedChunkSizeKB) << 10;
-    maxCodeLength = static_cast<length_type>(encodedMaxCodeLength);
-
-    return true;
-}
-
-//==============================================================================
-bool HuffmanCoder::encodeCodes(BitStreamWriter &output) const noexcept
+bool HuffmanEncoder::encodeCodes(BitStreamWriter &output) const noexcept
 {
     // At the least, encode that there were zero Huffman codes of length zero.
     std::vector<std::uint16_t> counts(1);
@@ -465,92 +382,7 @@ bool HuffmanCoder::encodeCodes(BitStreamWriter &output) const noexcept
 }
 
 //==============================================================================
-bool HuffmanCoder::decodeCodes(
-    BitStreamReader &input,
-    length_type &maxCodeLength) noexcept
-{
-    m_huffmanCodesSize = 0;
-
-    // Decode the number of code length counts. This number must be at least 1.
-    byte_type countsSize;
-    if (!input.ReadByte(countsSize) || (countsSize == 0))
-    {
-        return false;
-    }
-
-    // The first code length is 0, so the actual maximum code length is 1 less
-    // than the number of length counts.
-    maxCodeLength = countsSize - 1;
-
-    // Decode the code length counts.
-    std::vector<std::uint16_t> counts(countsSize);
-
-    for (std::uint16_t &count : counts)
-    {
-        if (!input.ReadWord(count))
-        {
-            return false;
-        }
-    }
-
-    // Decode the symbols.
-    for (length_type length = 0; length < countsSize; ++length)
-    {
-        for (std::uint16_t i = 0; i < counts[length]; ++i)
-        {
-            byte_type symbol;
-            if (!input.ReadByte(symbol))
-            {
-                return false;
-            }
-
-            // First code is always set to zero.
-            code_type code = 0;
-
-            // Subsequent codes are one greater than the previous code, but also
-            // bit-shifted left enough to maintain the right code length.
-            if (m_huffmanCodesSize != 0)
-            {
-                const HuffmanCode &last =
-                    m_huffmanCodes[m_huffmanCodesSize - 1];
-
-                const length_type shift = length - last.m_length;
-                code = (last.m_code + 1) << shift;
-            }
-
-            if (m_huffmanCodesSize == m_huffmanCodes.size())
-            {
-                return false;
-            }
-
-            m_huffmanCodes[m_huffmanCodesSize++] =
-                HuffmanCode(static_cast<symbol_type>(symbol), code, length);
-        }
-    }
-
-    convertToPrefixTable(maxCodeLength);
-    return true;
-}
-
-//==============================================================================
-void HuffmanCoder::convertToPrefixTable(length_type maxCodeLength) noexcept
-{
-    for (std::uint16_t i = 0; i < m_huffmanCodesSize; ++i)
-    {
-        const HuffmanCode code = std::move(m_huffmanCodes[i]);
-        const length_type shift = maxCodeLength - code.m_length;
-
-        for (code_type j = 0; j < (1_u16 << shift); ++j)
-        {
-            const code_type index = (code.m_code << shift) + j;
-            m_prefixTable[index].m_symbol = code.m_symbol;
-            m_prefixTable[index].m_length = code.m_length;
-        }
-    }
-}
-
-//==============================================================================
-bool HuffmanCoder::encodeSymbols(
+bool HuffmanEncoder::encodeSymbols(
     std::uint32_t inputSize,
     BitStreamWriter &output) noexcept
 {
@@ -574,34 +406,6 @@ bool HuffmanCoder::encodeSymbols(
     }
 
     return true;
-}
-
-//==============================================================================
-bool HuffmanCoder::decodeSymbols(
-    BitStreamReader &input,
-    length_type maxCodeLength,
-    std::uint32_t chunkSize,
-    std::ostream &output) const noexcept
-{
-    std::uint32_t bytes = 0;
-    code_type index;
-
-    while ((bytes < chunkSize) && (input.PeekBits(maxCodeLength, index) != 0))
-    {
-        const HuffmanCode &code = m_prefixTable[index];
-
-        m_chunkBuffer[bytes++] = code.m_symbol;
-        input.DiscardBits(code.m_length);
-    }
-
-    if (bytes > 0)
-    {
-        output.write(
-            reinterpret_cast<const std::ios::char_type *>(m_chunkBuffer.get()),
-            bytes);
-    }
-
-    return (bytes == chunkSize) || input.FullyConsumed();
 }
 
 } // namespace fly
