+++ fly/files.mk
@@ -1,6 +1,7 @@
 # Define the directories to include and compile
 SRC_DIRS_$(d) := \
     fly/coders \
+    fly/coders/huffman \
     fly/config \
     fly/logger \
     fly/parser \
